// src/api/generated/model/evaluationScope.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */

/**
 * Enumeration of what a contest evaluation run should cover.
 */
export type EvaluationScope = typeof EvaluationScope[keyof typeof EvaluationScope];


export const EvaluationScope = {
  ALL: 'ALL',
  TEAMS: 'TEAMS',
  QUESTIONS: 'QUESTIONS',
  STUDENTS: 'STUDENTS',
} as const;